Be more resilient by converting the array of SFR to float. This avoids a possible crash down the road when doing an in-place assignment of an integer with a float.

+++ pcigale/creation_modules/sfhfromfile.py
@@ -71,7 +71,7 @@ class SfhFromFile(CreationModule):
         time_grid = table.columns[0].data
 
         sfr_column_number = int(self.parameters['sfr_column'])
-        sfr = table.columns[sfr_column_number].data
+        sfr = table.columns[sfr_column_number].data.astype(np.float)
 
         age = int(self.parameters['age'])
         normalise = (self.parameters["normalise"].lower() == "true")
